Split into several files to make it easier to work with. (first part of the refactor task)
