+++ form.html
@@ -1,0 +1,30 @@
+<form method="post" enctype="multipart/form-data">
+<div class="panel panel-default">
+  <div class="panel-body">
+  <div class="form-group">
+          <label for="proxy">Proxy</label>
+          <input type="text" name="proxy" value=""/>
+</div>
+<div class="form-group">
+  <label for="username">Username</label>
+  <input type="text" name="username" value=""/>
+  <label for="password">Password</label>
+  <input type="pass" name="password" value=""/>
+</div>
+
+<div class="form-group">
+  <label for="find">Find</label>
+  <input type="text" name="find"/>
+  <label for="replace">Find</label>
+  <input type="text" name="replace"/>
+</div>
+<div class="form-group">
+          <label for="csv_input">Upload a CSV file</label>
+          <input type="file" name="csv_input" class="input-medium" required/>
+          <p>See <a href="example.csv">the example CSV file</a> for the expected format</p>
+          </div>
+
+        <input type="submit" class="btn"/>
+</div>
+</div>
+</form>
