+++ template_top.html
@@ -1,0 +1,26 @@
+<html>
+<head>
+  <title>URL Redirection Tester</title>
+  <link rel="stylesheet"
+        href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.2/css/bootstrap.min.css">
+  <link rel="stylesheet"
+        href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.2/css/bootstrap-theme.min.css">
+  <style>
+    tbody {
+      word-break: break-word;
+    }
+  </style>
+</head>
+<body>
+<a href="https://github.com/malcomio/redirect-tester">
+  <img style="position: absolute; top: 0; right: 0; border: 0;"
+       src="https://camo.githubusercontent.com/e7bbb0521b397edbd5fe43e7f760759336b5e05f/68747470733a2f2f73332e616d617a6f6e6177732e636f6d2f6769746875622f726962626f6e732f666f726b6d655f72696768745f677265656e5f3030373230302e706e67"
+       alt="Fork me on GitHub"
+       data-canonical-src="https://s3.amazonaws.com/github/ribbons/forkme_right_green_007200.png">
+</a>
+
+<div class="container">
+
+<h1>Redirection checker</h1>
+<p>Upload a CSV file containing original and expected redirected URLs to test whether your redirect rules are working as intended.</p>
+<p>See <a href="https://github.com/malcomio/redirect-tester/blob/master/README.md">the README</a> for more information.</p>
